Fix address text

+++ mobile_application/lib/models/eta.dart
@@ -3,8 +3,8 @@ import 'package:mobile_application/models/user.dart';
 class Eta {
   // final User driver;
   final Map<String, dynamic> driver;
-  final int etaStartBus;
-  final int etaEndBus;
+  final double etaStartBus;
+  final double etaEndBus;
   final DateTime timeOfArrival;
   final double distanceStartBus;
   final double distanceEndBus;
